// src/types/enums.ts
export enum THEME {
  White = "WHITE",
  Black = "BLACK",
}

export enum ConnectorNames {
  Injected = "injected",
  WalletConnect = "walletconnect",
  WalletLink = "walletlink",
  Fortmatic = "fortmatic",
}

export enum AssetType {
  SpotComposite = "Spot Composite",
}

export enum ETransactionItemType {
  Buy = "Bought",
  Sell = "Sell",
}